backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pyodbc==5.0.1
python-dotenv==1.0.0
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3
torch==2.1.2
azure-identity==1.15.0
gunicorn==21.2.0
